Create "FeedStoreSpecsProtocol" file that must be implemented by any "FeedStoreProtocol" test implementation to guarantee it meets spec
